+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The releases of Vieb aim to follow [semantic versioning](https://semver.org).
 - Settings of type list and object, that can also be set with lists or objects with ":set"
 - Setting "loadingindicator" to control where the page load status should be shown
 - Setting "passthroughkeys" to send keys to the page per site/domain/regex without leaving normal mode
+- Option to write command to save pages as MHTML instead of just HTML
 
 ## Changed